Update to fix issues #1 and #2
In this latest rollout of the Quotr app, issues #1 and #2 were both addressed and corrected. The app currently only supports portrait orientation which prevents users from rotating the screen in-app and causing a crash. Also, the app now determines the screen size of the device and fetches quotes based on their length and doesn't display one unless it meets the screen size and quote length criteria. App to be published to the Play Store soon!